abort if rowspan goes from thead to tbody: https://github.com/metanorma/metanorma-standoc/issues/648
